fixed bug : we can see old data in connection form when close and open the dailog again
