+++ src/main/webapp/app/modules/administration/sources/connections/connection-steps.reducer.ts
@@ -4,17 +4,12 @@ import { IDatasources, defaultDatasourceValue } from 'app/shared/model/datasourc
 import { IConnectionType, defaultConnectionTypeValue } from 'app/shared/model/connection-type.model';
 export const ACTION_TYPES = {
   SELECT_CONNECTION_TYPE: 'connectionSteps/SELECT_CONNECTION_TYPE',
-  SELECT_CONNECTION: 'connectionSteps/SELECT_CONNECTION',
-  SET_CONNECTION: 'connectionSteps/SET_CONNECTION',
   RESET: 'connectionSteps/RESET',
   IS_SAVE_CONNECTION: 'connectionSteps/IS_SAVE_CONNECTION',
-  IS_CONNECTION_SELECTED: 'connectionSteps/IS_CONNECTION_SELECTED',
 };
 
 const initialState = {
   connectionType: defaultConnectionTypeValue,
-  connection: connectionDefaultValue,
-  isConnectionSelected: false,
   isAddFeaturesCalled: false,
   isSaveConnectionCalled: false,
 };
@@ -27,22 +22,6 @@ export default (state: ConnectionStepsState = initialState, action): ConnectionS
       return {
         ...state,
         connectionType: action.payload,
-      };
-    case ACTION_TYPES.SELECT_CONNECTION:
-      return {
-        ...state,
-        connection: action.payload,
-        isConnectionSelected: true,
-      };
-    case ACTION_TYPES.IS_CONNECTION_SELECTED:
-      return {
-        ...state,
-        isConnectionSelected: action.payload,
-      };
-    case ACTION_TYPES.SET_CONNECTION:
-      return {
-        ...state,
-        connection: action.payload,
       };
     case ACTION_TYPES.IS_SAVE_CONNECTION:
       return {
@@ -64,26 +43,7 @@ export const selectConnectionType = (connectionType: IConnectionType) => {
     payload: connectionType,
   };
 };
-export const selectConnection = (connection: IConnection) => {
-  return {
-    type: ACTION_TYPES.SELECT_CONNECTION,
-    payload: connection,
-  };
-};
 
-export const setIsConnectionSelected = isConnectionSelected => {
-  return {
-    type: ACTION_TYPES.IS_CONNECTION_SELECTED,
-    payload: isConnectionSelected,
-  };
-};
-
-export const setConnection = (connection: IConnection) => {
-  return {
-    type: ACTION_TYPES.SET_CONNECTION,
-    payload: connection,
-  };
-};
 export const resetSteps = () => {
   return {
     type: ACTION_TYPES.RESET,
