+++ src/main/webapp/app/modules/administration/sources/connections/data-connection/data-connection.tsx
@@ -5,7 +5,7 @@ import { Button, Flex, Picker, Item, TextField, Form, ProgressBar, Text } from '
 import { getConnectionsByConnectionTypeId } from '../connection.reducer';
 import { prepareConnection, generateConnectionsOptions } from '../connections.util';
 import ConnectionProperty from './connection-property';
-import { selectConnection, setConnection, setIsConnectionSelected } from '../connection-steps.reducer';
+import { selectConnection, setConnection, setIsConnectionSelected } from '../connection.reducer';
 import { queryToConnection, setIsConnected } from '../../datasources/datasources.reducer';
 import { Translate, translate } from 'react-jhipster';
 import Checkmark from '@spectrum-icons/workflow/Checkmark';
@@ -141,8 +141,8 @@ export const DataConnection = (props: IDataConnectionProps) => {
 
 const mapStateToProps = (storeState: IRootState) => ({
   connections: storeState.connections.connections,
-  connection: storeState.connectionSteps.connection,
-  isConnectionSelected: storeState.connectionSteps.isConnectionSelected,
+  connection: storeState.connections.connection,
+  isConnectionSelected: storeState.connections.isConnectionSelected,
   isConnected: storeState.datasources.isConnected,
   errorMessage: storeState.datasources.errorMessage,
   loading: storeState.datasources.loading,
