+++ src/main/webapp/app/shared/model/connection.model.ts
@@ -1,14 +1,14 @@
 import { cacheDefaultValue, ICache } from './cache.model';
 export interface IConnection {
-  id: string;
-  name: string;
-  connectionUsername: string;
-  connectionPassword: string;
-  details: any;
-  connectionType: string;
-  connectionTypeId: number;
-  linkId: string;
-  connectionParameters: ICache;
+  readonly id: string;
+  readonly name: string;
+  readonly connectionUsername: string;
+  readonly connectionPassword: string;
+  readonly details: any;
+  readonly connectionType: string;
+  readonly connectionTypeId: number;
+  readonly linkId: string;
+  readonly connectionParameters: ICache;
 }
 
 export const connectionDefaultValue: Readonly<IConnection> = {
